+++ config/initializers/container.rb
@@ -259,6 +259,7 @@ module Charkeeper
       WebhooksContext::Telegram::HandleChatMemberWebhookService.new
     }
 
+    register('services.characters_context.pathfinder2.refresh_feats') { CharactersContext::Pathfinder2::RefreshFeats.new }
     register('services.characters_context.daggerheart.refresh_feats') { CharactersContext::Daggerheart::RefreshFeats.new }
     register('services.characters_context.dnd5.refresh_feats') { CharactersContext::Dnd5::RefreshFeats.new }
     register('services.characters_context.dnd2024.refresh_feats') { CharactersContext::Dnd2024::RefreshFeats.new }