+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - spells filtering for PF2 characters
 - focus and innate spells management
 - pet/familiar management for PF2 characters
+- static feats for PF2 characters
 
 ### Modified
 - builders for PF2 characters